+++ src/Areas/DataFactory/Commands/BaseDataFactoryCommand.cs
@@ -1,0 +1,30 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+using System.Diagnostics.CodeAnalysis;
+using AzureMcp.Areas.DataFactory.Options;
+using AzureMcp.Commands;
+using AzureMcp.Commands.Subscription;
+
+namespace AzureMcp.Areas.DataFactory.Commands;
+
+public abstract class BaseDataFactoryCommand<
+    [DynamicallyAccessedMembers(TrimAnnotations.CommandAnnotations)] T>
+    : SubscriptionCommand<T>
+    where T : BaseDataFactoryOptions, new()
+{
+    protected readonly Option<string> _factoryOption = DataFactoryOptionDefinitions.Factory;
+
+    protected override void RegisterOptions(Command command)
+    {
+        base.RegisterOptions(command);
+        command.AddOption(_factoryOption);
+    }
+
+    protected override T BindOptions(ParseResult parseResult)
+    {
+        var options = base.BindOptions(parseResult);
+        options.FactoryName = parseResult.GetValueForOption(_factoryOption);
+        return options;
+    }
+}